Add travis patch testing.
This adds a .yml file and a bunch of utility scripts
that apply the gentoo patch series against a specified
upstream tar.bz2, and then compiles and tests it while mimicing
a large set of Gentoo Configuration.

Tests on Native GCC ( currently 4.6 ), GCC5, GCC6
and CLANG (3.4).

Logic is present to test on CLANG3.8, however, upstreams
mirroring of packages for CLANG3.8 has been broken due to traffic
load and may need fixing, so is marked an "allowed failure".

One patch doesn't appear to work on the travis target due to
something weird going on with -Duseshrplib, and so that
patch is excluded from the series.

Only one non-Gentoo configure flag passed is -Dusedevel, which
allows testing against 5.25* if the TARGET is changed.

However, this needs to be looked into and worked out why it fails,
because it likely indicates a failure somewhere else.